+++ src/main/java/net/optifine/util/MathUtils.java
@@ -11,12 +11,12 @@ public class MathUtils
 
     public static float asin(float value)
     {
-        return ASIN_TABLE[(int)((value + 1.0F) * 32767.5D) & 65535];
+        return ASIN_TABLE[(int)((double)(value + 1.0F) * 32767.5D) & 65535];
     }
 
     public static float acos(float value)
     {
-        return ((float)Math.PI / 2F) - ASIN_TABLE[(int)((value + 1.0F) * 32767.5D) & 65535];
+        return ((float)Math.PI / 2F) - ASIN_TABLE[(int)((double)(value + 1.0F) * 32767.5D) & 65535];
     }
 
     public static int getAverage(int[] vals)
@@ -76,19 +76,19 @@ public class MathUtils
 
     public static float roundToFloat(double d)
     {
-        return (float)(Math.round(d * 1.0E8D) / 1.0E8D);
+        return (float)((double)Math.round(d * 1.0E8D) / 1.0E8D);
     }
 
     static
     {
         for (int i = 0; i < 65536; ++i)
         {
-            ASIN_TABLE[i] = (float)Math.asin(i / 32767.5D - 1.0D);
+            ASIN_TABLE[i] = (float)Math.asin((double)i / 32767.5D - 1.0D);
         }
 
         for (int j = -1; j < 2; ++j)
         {
-            ASIN_TABLE[(int)((j + 1.0D) * 32767.5D) & 65535] = (float)Math.asin(j);
+            ASIN_TABLE[(int)(((double)j + 1.0D) * 32767.5D) & 65535] = (float)Math.asin((double)j);
         }
     }
 }
